Moving `handleFocalpoint()` to decorate.js
