+++ libs/utils/decorate.js
@@ -75,6 +75,24 @@ export function decorateBlockText(el, config = ['m', 's', 'm'], type = null) {
   if (type === 'merch') decorateIconStack(el);
 }
 
+export function handleFocalpoint(pic, child, removeChild) {
+  const image = pic.querySelector('img');
+  if (!child || !image) return;
+  let text = '';
+  if (child.childElementCount === 2) {
+    const dataElement = child.querySelectorAll('p')[1];
+    text = dataElement?.textContent;
+    if (removeChild) dataElement?.remove();
+  } else if (child.textContent) {
+    text = child.textContent;
+    const childData = child.childNodes;
+    if (removeChild) childData.forEach((c) => c.nodeType === Node.TEXT_NODE && c.remove());
+  }
+  const directions = text.trim().toLowerCase().split(',');
+  const [x, y = ''] = directions;
+  image.style.objectPosition = `${x} ${y}`;
+}
+
 export async function decorateBlockBg(block, node) {
   const childCount = node.childElementCount;
   if (node.querySelector('img, video, a[href*=".mp4"]') || childCount > 1) {
@@ -82,16 +100,15 @@ export async function decorateBlockBg(block, node) {
     const binaryVP = [['mobile-only'], ['tablet-only', 'desktop-only']];
     const allVP = [['mobile-only'], ['tablet-only'], ['desktop-only']];
     const viewports = childCount === 2 ? binaryVP : allVP;
-    const { handleFocalpoint } = await import('../blocks/section-metadata/section-metadata.js');
     [...node.children].forEach((child, i) => {
       const videoLink = child.querySelector('a[href*=".mp4"]');
-      const pic = child.querySelector('picture');
-      if (childCount > 1) child.classList.add(...viewports[i]);
       if (videoLink && !videoLink.hash) videoLink.hash = 'autoplay';
+      if (childCount > 1) child.classList.add(...viewports[i]);
+      const pic = child.querySelector('picture');
       if (pic && (child.childElementCount === 2 || child.textContent?.trim())) {
         handleFocalpoint(pic, child, true);
       }
-      if (!pic && !videoLink) {
+      if (!child.querySelector('img, video, a[href*=".mp4"]')) {
         child.style.background = child.textContent;
         child.textContent = '';
       }
